Continue of major refactoring of TM_Xml_Database   - For example now the TM_Xml_Database ctor does not trigger a full data load (when in UsingFileSystem mode)   - TM data loaded is now mostly at TM_XmlDatabase/TM_Xml_Database.Config.cs This version does not load all libraries (but the main UI and Userdata should load) Removed TMConfig.config file form the web root Removed quite a bit of code (for features that are not neeed anymore)   - For example the UserData 'FirstScript to invoke' feature Refactored a number of string values into the TMConsts.cs class Fixed ContextType bug in 'TM is Unavailable' page Added a new type of Events class for the TM_Xml_Database Lots of new tests to cover changes made (note that are 57 currently breaking (due to the fact that libraries are not being loaded)

+++ Web Applications/TeamMentor.UnitTests/_Test_ConfigAndHelpers/Misc_CoreLib_ExtensionMethods.cs
@@ -17,20 +17,27 @@ namespace TeamMentor.UnitTests
             Assert.NotNull    (tmDatabase);
             Assert.NotNull    (tmDatabase.Path_XmlDatabase);
             Assert.AreNotEqual(tmDatabase.Path_XmlDatabase, TM_Server.WebRoot);
-            Assert.IsNotEmpty (tmDatabase.Path_XmlDatabase.files());
-            Assert.IsTrue     (tmDatabase.Path_XmlDatabase.dirExists());
-            //make all files writable
-            tmDatabase.Path_XmlDatabase.files(true).files_Attribute_ReadOnly_Remove();
-            
-            //delete all files recusively
-            Files.deleteFolder(tmDatabase.Path_XmlDatabase, true);
+            //Assert.IsTrue     (tmDatabase.Path_XmlDatabase.dirExists());
 
-            //ensure the deletion happened
-            Assert.IsFalse    (tmDatabase.Path_XmlDatabase.dirExists());
-            Assert.IsEmpty    (tmDatabase.Path_XmlDatabase.files());
+            if (tmDatabase.Path_XmlDatabase.dirExists())                                            // check if the folder exists      
+            {
+                //Assert.IsNotEmpty(tmDatabase.Path_XmlDatabase.files());
 
-            "[Test][TM_Xml_Database][delete_Database]TM database files were deleted from: {0}".info(tmDatabase.Path_XmlDatabase);
+                tmDatabase.Path_XmlDatabase.files(true).files_Attribute_ReadOnly_Remove();          // make all files writable
+
+                Files.deleteFolder(tmDatabase.Path_XmlDatabase, true);                              // delete all files recusively
+
+                Assert.IsFalse(tmDatabase.Path_XmlDatabase.dirExists());                            // ensure the deletion happened
+                Assert.IsEmpty(tmDatabase.Path_XmlDatabase.files());
+
+                "[Test][TM_Xml_Database][delete_Database]TM database files were deleted from: {0}".info(tmDatabase.Path_XmlDatabase);
+            }
             return tmDatabase;
+        }        
+        public static TM_Xml_Database loadData_No_Admin_User(this TM_Xml_Database tmDatabase)
+        {
+            tmDatabase.Server.Users_Create_Default_Admin = false;
+            return tmDatabase.loadData();
         }
     }
 }
